+++ src/app/app-routing.module.ts
@@ -1,16 +1,17 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { CartListComponent } from './cart/components/cart-list/cart-list.component';
+import { FirstComponent } from './first/components/first/first.component';
 import { ProductListComponent } from './products/components/product-list/product-list.component';
 
 const routes: Routes = [
   {
-    path: 'products-list',
-    component: ProductListComponent
-  },
-  {
     path: 'cart',
     component: CartListComponent
+  },
+  {
+    path: 'first',
+    component: FirstComponent
   },
   {
     path: '',