+++ src/app/app.module.ts
@@ -6,6 +6,7 @@ import { AppComponent } from './app.component';
 import { FirstModule } from './first/first.module';
 import { ProductsModule } from './products/products.module';
 import { CartModule } from './cart/cart.module';
+// import { Router } from '@angular/router';
 
 @NgModule({
   declarations: [
@@ -21,4 +22,11 @@ import { CartModule } from './cart/cart.module';
   providers: [],
   bootstrap: [AppComponent]
 })
-export class AppModule { }
+export class AppModule {
+  // constructor(router: Router) {
+  //   const replacer = (key: string, value: any): string =>
+  //     typeof value === 'function' ? value.name : value;
+
+  //   console.log('Routes: ', JSON.stringify(router.config, replacer, 2));
+  // }
+}