+++ src/app/products/components/product-view/product-view.component.ts
@@ -1,0 +1,31 @@
+import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute, ParamMap, Router } from '@angular/router';
+import { Product } from '../../models/product.model';
+import { ProductsService } from '../../services/products.service';
+import { switchMap } from 'rxjs/operators';
+
+@Component({
+  selector: 'app-product-view',
+  templateUrl: './product-view.component.html',
+  styleUrls: ['./product-view.component.scss']
+})
+export class ProductViewComponent implements OnInit {
+  product: Product = null;
+
+  constructor(
+    private productsService: ProductsService,
+    private router: Router,
+    private route: ActivatedRoute
+  ) { }
+
+  ngOnInit(): void {
+    const observer = {
+      next: (product: Product) => (this.product = { ...product }),
+      error: (err: any) => console.log(err)
+    };
+    this.route.paramMap
+      .pipe(
+        switchMap((params: ParamMap) => this.productsService.getProductById(`${params.get('productId')}`)))
+      .subscribe(observer);
+  }
+}