+++ src/app/app-routing.module.ts
@@ -1,13 +1,14 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { CartListComponent } from './cart/components/cart-list/cart-list.component';
+import { IsCartEmptyGuard } from './core/guards/is-cart-empty.guard';
 import { FirstComponent } from './first/components/first/first.component';
 import { ProductListComponent } from './products/components/product-list/product-list.component';
 
 const routes: Routes = [
   {
-    path: 'cart',
-    component: CartListComponent
+    path: 'order',
+    canLoad: [IsCartEmptyGuard],
+    loadChildren: () => import('./order/order.module').then(m => m.OrderModule),
   },
   {
     path: 'first',