+++ src/app/cart/cart.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { CartListComponent } from './components/cart-list/cart-list.component';
 import { CartItemComponent } from './components/cart-item/cart-item.component';
 import { SharedModule } from '../shared/shared.module';
+import { CartRoutingModule } from './cart-routing.module';
 
 @NgModule({
   declarations: [
@@ -10,7 +10,8 @@ import { SharedModule } from '../shared/shared.module';
     CartItemComponent
   ],
   imports: [
-    SharedModule
+    SharedModule,
+    CartRoutingModule
   ],
   exports: [
     CartListComponent